Begin move to generalized Project/Goals/Barriers format
- moved existing working model to Habitat_Opt_orig.gms for referencing
while creating new code
- created new code template and test file Habitat_Opt_test.gms which
contains the new model format as well as a small test dataset for
testing if the model works properly. This version, instead of using
Barriers, RestorationFishes, ControlFishes, has moved to a more general
framework of Barriers, Projects, Goals, where at each barrier a decision
is made to do a project that has some contribution on a goal at that
barrier. Projects contribute to a goal by increasing the delivery
efficiency of that goal at the barrier, i.e. the total delivery
efficiency of a goal at a barrier is a function of a baseline
efficiency, the total efficiency from downstream barriers, and the
efficiency increase from doing a project at that barrier. This
efficiency is multiplied by a maximum possible benefit to a goal at a
barrier to get the total benefit to a goal at a barrier. This version
also uses a variation on the Root barriers that reduces the number of
model equations while increasing the number of barriers by 1. A dummy
barrier is defined that is the downstream barrier of all root barriers.
This dummy barrier always has a total efficiency for all projects of 1,
effectively allowing all Root barriers to be treated as any other
barrier.
- initial tests indicate the new version works as planned. Future
versions will have to be tested on actual data