changes to definitions.csv syntax, additional error checking
- changed most occurrences of 'parameter' to 'symbol' to match with GAMS
convention and disambiguate two types of information (i.e. Parameters is
a subclass of symbols)
- Added a column 'Values is Column Name' requirement to definitions.csv
that indicates whether the value in the 'Values' column indicates a
column name in table.csv or is a (series of) value(s) in
definitions.csv. Additional checking now occurs in make_gdx.py to ensure
users specify this in addition to using the default method of taking all
'Barrires' indexed symbols from table.csv.
- changed the way the "default run" is defined in definitions.csv and
processed in make_gdx.py. Now a blank "run" indicates a default value,
and if no default values are given, it is assumed the default values
come from the lowest-number run.  The default values are not used to
make a separate run GDX
- updated various comments
- added a check to see if a user specifies multiple rows with the same
run + symbol + indices combination, in which case an error is raised
- added a check to see that indices specified in the Symbol column are
elements of Sets defined elsewhere in the input data. This check occurs
in make_gdx()